my-yolov8-app/backend/requirements.txt
Flask==3.0.2
flask-cors==4.0.1
Pillow==10.3.0
ultralytics==8.2.0
torch==2.2.1
torchvision==0.17.1
gunicorn==21.2.0
python-dotenv==1.0.1
pytest==8.2.0
